fix(secrets): support daemonless CLI
Extract the encrypted local store and execution primitives into @signet/core so local CLI commands can operate without the daemon. Preserve the daemon HTTP queue and external provider boundary, and coordinate store writers with an atomic process lock.

Closes #1593
Related: #1454
