+++ platform/core/package.json
@@ -37,6 +37,8 @@
 		"typecheck": "tsc --noEmit"
 	},
 	"dependencies": {
+		"@napi-rs/keyring": "1.3.0",
+		"libsodium-wrappers": "^0.8.2",
 		"sqlite-vec": "^0.1.7-alpha.2",
 		"yaml": "^2.6.0"
 	},
@@ -50,6 +52,7 @@
 	},
 	"devDependencies": {
 		"@types/better-sqlite3": "^7.6.0",
+		"@types/libsodium-wrappers": "^0.8.2",
 		"@types/node": "^22.0.0",
 		"typescript": "^5.7.0",
 		"@signet/native": "workspace:*"
