+++ platform/core/src/secrets-keyring.ts
